[librpbase] FileSystem_posix.cpp: Cast sf_buf.f_type to uint32_t.
I encountered an error while trying to build on 32-bit Linux using
clang-9.0 for Address Sanitizer, though it probably happens on 64-bit
Linux with clang as well.

This fixes several compile errors that look like this:

src/librpbase/file/FileSystem_posix.cpp:354:8: error: case value
evaluates to 3730735588, which cannot be narrowed to type '__fsword_t'
(aka 'int') [-Wc++11-narrowing]
                case EFIVARFS_MAGIC:
                     ^
/usr/include/linux/magic.h:33:25: note: expanded from macro 'EFIVARFS_MAGIC'
.#define EFIVARFS_MAGIC         0xde5e81e4
                                ^

+++ src/librpbase/file/FileSystem_posix.cpp
@@ -348,7 +348,7 @@ bool isOnBadFS(const char *filename, bool netFS)
 		return false;
 	}
 
-	switch (sfbuf.f_type) {
+	switch (static_cast<uint32_t>(sfbuf.f_type)) {
 		case DEBUGFS_MAGIC:
 		case DEVPTS_SUPER_MAGIC:
 		case EFIVARFS_MAGIC:
